+++ packages/cli/generate-api/src/generator/generator.ts
@@ -1,12 +1,10 @@
-import {
-  generateApi,
-  GenerateApiOutput,
-  GenerateApiParams,
-} from "swagger-typescript-api";
+import { generateApi } from "swagger-typescript-api";
 import path from "node:path";
 import fs from "node:fs";
 
-export const generateSwaggerApi = async (params: GenerateApiParams) => {
+export const generateSwaggerApi = async (
+  params: Parameters<typeof generateApi>[0]
+): ReturnType<typeof generateApi> => {
   return await generateApi(params)
     .then((generateApiOutput) => {
       return generateApiOutput;
@@ -17,17 +15,34 @@ export const generateSwaggerApi = async (params: GenerateApiParams) => {
     });
 };
 
-export const writeGeneratedApi = async (
-  result: GenerateApiOutput,
-  outputPath: string
-) => {
-  const { files } = result;
+export const writeGeneratedApi = async ({
+  result,
+  outputPath,
+  httpClientRewrite,
+}: {
+  result: Awaited<ReturnType<typeof generateApi>>;
+  outputPath: string;
+  httpClientRewrite?: boolean;
+}) => {
+  const { files } = await result;
   files.forEach((element) => {
     const { fileName, fileContent } = element;
+
+    console.log(`📄 ${fileName} 파일 내용 길이:`, fileContent.length);
+    console.log(
+      `📄 ${fileName} 파일 내용 미리보기:`,
+      fileContent.substring(0, 200)
+    );
+
     const folderPath = getFolderPath(outputPath, fileName);
 
     fs.mkdirSync(folderPath, { recursive: true });
-    if (fileName === "http-client" || fileName === "data-contracts") {
+    if (fileName === "http-client") {
+      if (httpClientRewrite) {
+        createFile(path.resolve(folderPath, "index.ts"), fileContent);
+      }
+      return;
+    } else if (fileName === "data-contracts") {
       createFile(path.resolve(folderPath, "index.ts"), fileContent);
       return;
     }
@@ -59,7 +74,7 @@ const commentTemplate = `/* eslint-disable */
  * ---------------------------------------------------------------
  * ## THIS FILE WAS GENERATED VIA SWAGGER-TYPESCRIPT-API        ##
  * ##                                                           ##
- * ## AUTHOR: acacode                                           ##
+ * ## AUTHOR: brightbong92                                      ##
  * ## SOURCE: https://github.com/acacode/swagger-typescript-api ##
  * ---------------------------------------------------------------
  */