+++ packages/cli/generate-api/src/index.ts
@@ -4,15 +4,16 @@ import {
   generateSwaggerApi,
   writeGeneratedApi,
 } from "@/generator/generator.js";
-import { GenerateApiParams } from "swagger-typescript-api";
+import { generateApi } from "swagger-typescript-api";
 import { generateConfig } from "./configs/generate-config.js";
 
 export type GenerateApiParamsType = {
+  httpClientRewrite?: boolean;
   httpClientType: "fetch" | "axios";
   instancePath?: string;
   url: string;
   output: string;
-} & GenerateApiParams;
+} & Parameters<typeof generateApi>[0];
 
 export const runGenerateApi = async (params: GenerateApiParamsType) => {
   const { instancePath, output } = params || {};
@@ -21,9 +22,10 @@ export const runGenerateApi = async (params: GenerateApiParamsType) => {
     typeSuffix: "Type",
     addReadonly: true,
     output: false,
-    modular: true,
     moduleNameFirstTag: false,
     silent: true,
+    modular: true,
+    httpClientType: params.httpClientType,
     templates:
       params.httpClientType === "axios"
         ? generateConfig["CUSTOM_TEMPLATE_AXIOS"]
@@ -40,7 +42,11 @@ export const runGenerateApi = async (params: GenerateApiParamsType) => {
     },
   });
 
-  writeGeneratedApi(result, path.resolve(process.cwd(), output))
+  await writeGeneratedApi({
+    result,
+    outputPath: path.resolve(process.cwd(), output),
+    httpClientRewrite: params.httpClientRewrite,
+  })
     .then(() => {
       console.log("✅ API 생성 완료! 🌈✨");
       process.exit(0);