+++ apps/my-playground/CHANGELOG.md
@@ -1,5 +1,56 @@
 # my-playground
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @mash-up-web-toolkit/command@0.0.14
+
+## 0.1.8
+
+### Patch Changes
+
+- @mash-up-web-toolkit/command@0.0.13
+
+## 0.1.7
+
+### Patch Changes
+
+- @mash-up-web-toolkit/command@0.0.12
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @mash-up-web-toolkit/command@0.0.11
+
+## 0.1.5
+
+### Patch Changes
+
+- @mash-up-web-toolkit/command@0.0.10
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @mash-up-web-toolkit/command@0.0.9
+
+## 0.1.3
+
+### Patch Changes
+
+- @mash-up-web-toolkit/command@0.0.8
+
+## 0.1.2
+
+### Patch Changes
+
+- @mash-up-web-toolkit/command@0.0.7
+
 ## 0.1.1
 
 ### Patch Changes