+++ packages/cli/command/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @mash-up-web-toolkit/command
 
+## 0.0.14
+
+### Patch Changes
+
+- writeGeneratedApi 함수에 httpClientRewrite 옵션을 추가했습니다.
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-config@0.0.7
+  - @mash-up-web-toolkit/generate-api@0.0.12
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-api@0.0.11
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-api@0.0.10
+
+## 0.0.11
+
+### Patch Changes
+
+- 설정 파일 로드 시 동적 모듈 형식 감지 기능 추가
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-config@0.0.6
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-config@0.0.5
+
+## 0.0.9
+
+### Patch Changes
+
+- fix axios instance
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-api@0.0.9
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-api@0.0.8
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [6bd6308]
+  - @mash-up-web-toolkit/generate-api@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes